daemon: Prevent new daemon created on same machine
.bosch.com>

+++ src/daemon/dlt_daemon_unix_socket.c
@@ -145,7 +145,7 @@ int dlt_daemon_unix_socket_open(int *sock, char *sock_path, int type, int mask)
     addr.sun_family = AF_UNIX;
     memcpy(addr.sun_path, sock_path, sizeof(addr.sun_path));
 
-    if (unlink(sock_path) != 0) {
+    if (unlink(sock_path) != 0 && (errno != ENOENT)) {
         dlt_vlog(LOG_WARNING, "%s: unlink() failed: %s\n",
                 __func__, strerror(errno));
     }
